added empId field in both retail and textile product class

+++ inventory-module/src/main/java/com/otsi/retail/inventory/controller/ProductItemController.java
@@ -60,7 +60,7 @@ public class ProductItemController {
 	}
 
 	@GetMapping("/getBarcodeId")
-	public GateWayResponse<?> getBarcodeId(@RequestParam("barcodeId") Long barcodeId) {
+	public GateWayResponse<?> getBarcodeId(@RequestParam("barcodeId") int barcodeId) {
 		log.info("Recieved request to getBarcodeId:" + barcodeId);
 		ProductItemVo inventory = productItemService.getBarcodeId(barcodeId);
 		return new GateWayResponse<>("fetching barcode details successfully with barcodeId", inventory);
